AuthPost

+++ app/Http/Controllers/PostController.php
@@ -4,8 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Models\Tag;
 use App\Models\Post;
-use Illuminate\Http\Request;
+use App\Http\Requests\PostRequest;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -35,7 +36,7 @@ class PostController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(PostRequest $request)
     {
         $post = new Post();
         $post->title = $request->title;
@@ -44,6 +45,7 @@ class PostController extends Controller
         $post->publication_date = $request->publication_date;
         $post->user_id = Auth::user()->id;
         $post->save();
+        return redirect()->route('posts.index')->with('message', 'Post creato correttamente!');
 
         // salvataggio infoPost
         // $data["post_id"] = $post->id; //devo specificare il nuovo post_id con l'id del post creato
@@ -58,7 +60,7 @@ class PostController extends Controller
         //     }
         // }
 
-        return redirect()->route('posts.index')->with('message', 'Post creato correttamente!');
+
     }
 
     /**
@@ -66,6 +68,7 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+
         $tags = Tag::all();
         return view('posts.show', compact('post', 'tags'));
     }
@@ -75,6 +78,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+
         $tags = Tag::all();
         return view('posts.edit', compact('post', 'tags'));
     }
@@ -82,8 +86,11 @@ class PostController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Post $post)
+    public function update(PostRequest $request, Post $post)
     {
+        if (Gate::denies('edit-post', $post)) {
+            abort(403); // Retorna un error 403 (Forbidden) si no tiene permiso
+        }
         // modifica post
         $data = $request->all();
         $post->update($data);
@@ -110,6 +117,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if (Gate::denies('delete-post', $post)) {
+            abort(403); // Retorna un error 403 (Forbidden) si no tiene permiso
+        }
         $post->delete();
         return redirect()->route('posts.index')->with('message', 'Post elminato correttamente!');
     }
